requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
click==8.1.7
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
msgpack==1.0.8
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
tarantool==1.2.0
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.6
